requirements.txt
requests==2.32.3
python-dotenv==1.1.0
pyTelegramBotAPI==4.27.0
stravalib==2.2
schedule==1.2.2
google-genai==1.5.0